+++ FasterRCNN/FasterRCNN_Video.py
@@ -75,7 +75,7 @@ def getBoundingBoxAndCentroid(image, boxes, labels, scores, score_threshold=0.5)
     # boxes /= ratio
     output = image.copy()
     centroid = np.array([0,0])
-    
+    boxes = boxes.astype(np.int32)
     # Showing boxes with score > threshold
     for box, label, score in zip(boxes, labels, scores):
         if score > score_threshold and label == 37:  # Sports Ball label = 37
